CleanArchitectureDemo-31 Extract `LoggingInterceptor` to separate `debug` build module
