+++ data/src/release/kotlin/com/criticalgnome/data/di/OkhttpClientModule.kt
@@ -1,0 +1,19 @@
+package com.criticalgnome.data.di
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import okhttp3.OkHttpClient
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object OkhttpClientModule {
+
+    @Provides
+    @Singleton
+    fun provideOkHttpClient(): OkHttpClient {
+        return OkHttpClient()
+    }
+}
